Built the basic code for the project view routing
Project_view.html will route to a different view depending on the role
of the current user in this project. The controller in Project_view
will decide which template to load, and pass the project number to the
new template (currently sandbox.html, a html doc for testing purposes)
so that the controller in this new template will get the relevant data
from this project.